disables multiple operations on content simultaneously
