+++ src/components/ucp-tabs/content-manager/buttons/uninstall-button.tsx
@@ -1,16 +1,14 @@
-import { atom, useAtomValue, useSetAtom } from 'jotai';
+import { useAtomValue, useSetAtom } from 'jotai';
 import { XSquare } from 'react-bootstrap-icons';
-import { useMemo } from 'react';
 import { STATUS_BAR_MESSAGE_ATOM } from '../../../footer/footer';
 import {
+  BUSY_CONTENT_ELEMENTS_ATOM,
   CONTENT_INTERFACE_STATE_ATOM,
-  contentInstallationStatusAtoms,
 } from '../state/atoms';
 import { showModalOk } from '../../../modals/modal-ok';
 import { uninstallContents } from './callbacks/uninstall-content';
 import { showModalOkCancel } from '../../../modals/modal-ok-cancel';
 import Logger from '../../../../util/scripts/logging';
-import { getStore } from '../../../../hooks/jotai/base';
 
 const LOGGER = new Logger('uninstall-button.tsx');
 
@@ -21,17 +19,6 @@ export function UninstallButton(
   const setStatusBarMessage = useSetAtom(STATUS_BAR_MESSAGE_ATOM);
 
   const interfaceState = useAtomValue(CONTENT_INTERFACE_STATE_ATOM);
-
-  const selectedIDsAtom = useMemo(
-    () =>
-      atom(() =>
-        interfaceState.selected.map(
-          (ce) => `${ce.definition.name}@${ce.definition.version}`,
-        ),
-      ),
-    [interfaceState.selected],
-  );
-  const selectedIDs = useAtomValue(selectedIDsAtom);
   const selectionCount = interfaceState.selected.length;
   const installedCount = interfaceState.selected.filter(
     (ce) => ce.installed,
@@ -42,10 +29,8 @@ export function UninstallButton(
   const deprecated = interfaceState.selected.filter(
     (ce) => ce.installed && !ce.online,
   );
-  const busy = selectedIDs
-    .map((id) => getStore().get(contentInstallationStatusAtoms(id)))
-    .filter((cis) => cis.action !== 'idle');
-  const busyCount = busy.length;
+
+  const busyCount = useAtomValue(BUSY_CONTENT_ELEMENTS_ATOM).length;
 
   const countElement =
     selectionCount < 2 ? (
